requirements_realtime.txt
# Real-Time Trading System Requirements v2
# Core async libraries
asyncio-mqtt==0.13.0
aiohttp==3.9.1
aiofiles==23.2.1

# Supabase integration
supabase==2.3.0
postgrest==0.13.0
gotrue==2.12.0
realtime==1.0.0

# Database and ORM
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# MetaTrader 5 integration
MetaTrader5==5.0.5050

# Data processing
numpy==1.24.3
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.2.0
colorama==0.4.6

# Type hints
typing-extensions==4.8.0

# System monitoring and management
psutil==5.9.5

# Task management and concurrency
asyncio-throttle==1.0.2

# JSON handling for events
orjson==3.9.10

# WebSocket support for real-time connections
websockets==12.0

# HTTP client for enhanced requests
httpx==0.25.2

# Configuration management
pyyaml==6.0.1

# Enhanced asyncio utilities
anyio==4.2.0

# Testing (optional)
pytest-asyncio==0.23.2
pytest==7.4.3

# Development (optional)
black==23.12.0
isort==5.13.2
mypy==1.8.0 